Important travail d'automatisation et de simplification des classes et des initialisations. (la classe Info est un workaround pour un petit soucis de droit de modification, rien ne change pour vous). Modification visible depuis l'interface :  - Les inficateurs affichent leurs valeurs dès le lancement.  - Ajout d'un bouton "next turn" qui joue un tour. - Jouer un tours modifie la valeur de l'indicateur food Global. - Les indicateurs s'update maintenant selon leurs constantes.
