+++ Assets/Scripts/Classes/Info.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 34d142d2e94d3da4cac3134513b369e2
+timeCreated: 1450309762
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
