+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 5.2.3f1
+m_EditorVersion: 5.2.2f1
 m_StandardAssetsVersion: 0
